+++ examples/counter/test/widget_test.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test('testWidgets', () {
+    expect(2 + 3, 5);
+  });
+}