+++ packages/flutterfire_gen/lib/src/templates/query/doc_comment_template.dart
@@ -26,22 +26,27 @@ class QueryClassDocCommentTemplate {
 ///
 /// This class provides methods to perform CRUD (Create, Read, Update, Delete)
 /// operations on ${config.documentName} documents, along with additional utilities like counting
-/// documents.
+/// documents, and calculating sum and average values for specific fields.
 ///
-/// It includes methods to fetch, subscribe to, and count single or multiple [${config.readClassName}]
-/// documents, as well as methods to add, set, update, and delete documents.
+/// It includes methods to:
+///
+/// - Fetch single or multiple [${config.readClassName}] documents ([fetchDocuments], [fetchDocument]).
+/// - Subscribe to real-time updates of single or multiple [${config.readClassName}] documents ([subscribeDocuments], [subscribeDocument]).
+/// - Count documents based on queries ([count]).
+/// - Calculate sum ([getSum]) and average ([getAverage]) of specific fields across documents.
+/// - Add ([add]), set ([set]), update ([update]), and delete ([delete]) ${config.documentName} documents.
 ///
 /// The class uses Firebase Firestore as the backend, assuming [${config.readClassName}],
 /// [${config.createClassName}], [${config.updateClassName}] are models representing the data.
 ///
 /// Usage:
 ///
-/// - To fetch, subscribe to, or count one or more ${config.documentName} documents, use
-/// [fetchDocuments], [subscribeDocuments], [fetchDocument], [subscribeDocument], or [count].
-/// - To modify ${config.documentName} documents, use [add], [set], [update], or [delete].
+/// - To fetch or subscribe to ${config.documentName} documents, or to count them, use the corresponding fetch, subscribe, and count methods.
+/// - To modify ${config.documentName} documents, use the methods for adding, setting, updating, or deleting.
+/// - To perform aggregate calculations like sum and average, use [getSum] and [getAverage].
 ///
-/// This class is designed to abstract the complexities of direct Firestore
-/// usage and provide a straightforward API for ${config.documentName} document operations.
+/// This class abstracts the complexities of direct Firestore usage and provides
+/// a straightforward API for ${config.documentName} document operations.
 ''';
   }
 
@@ -59,7 +64,7 @@ class QueryClassDocCommentTemplate {
 ///
 /// Parameters:
 ///
-/// - [options] Optional `GetOptions` to define the source of the documents (server, cache).
+/// - [options] Optional [GetOptions] to define the source of the documents (server, cache).
 /// - [queryBuilder] Optional function to build and customize the Firestore query.
 /// - [compare] Optional function to sort the ${config.readClassName} documents.
 /// - [asCollectionGroup] Fetch the '${config.collectionName}' as a collection group if true.
@@ -109,6 +114,48 @@ class QueryClassDocCommentTemplate {
 /// - [asCollectionGroup] Count the '${config.collectionName}' as a collection group if true.
 ///
 /// Returns the count of documents as an integer.
+''';
+  }
+
+  /// Returns the doc comment for the getSum method.
+  String forGetSumMethod() {
+    return '''
+/// Returns the sum of the values of the documents that match the query.
+/// 
+/// This method returns the sum of the values of the documents that match the query.
+/// You can customize the query by using the [queryBuilder].
+/// The [asCollectionGroup] parameter determines whether to query the '${config.collectionName}'
+/// collection directly (false) or as a collection group across different
+/// Firestore paths (true).
+/// 
+/// Parameters:
+/// 
+/// - [field] The field to sum over.
+/// - [queryBuilder] Optional function to build and customize the Firestore query.
+/// - [asCollectionGroup] Query the '${config.collectionName}' as a collection group if true.
+/// 
+/// Returns the sum of the values of the documents that match the query.
+''';
+  }
+
+  /// Returns the doc comment for the getAverage method.
+  String forGetAverageMethod() {
+    return '''
+/// Returns the average of the values of the documents that match the query.
+/// 
+/// This method returns the average of the values of the documents that match the query.
+/// You can customize the query by using the [queryBuilder].
+/// The [asCollectionGroup] parameter determines whether to query the '${config.collectionName}'
+/// collection directly (false) or as a collection group across different
+/// Firestore paths (true).
+/// 
+/// Parameters:
+/// 
+/// - [field] The field to average over.
+/// - [queryBuilder] Optional function to build and customize the Firestore query.
+/// - [asCollectionGroup] Query the '${config.collectionName}' as a collection group if true.
+/// 
+/// Returns the average of the values of the documents that match the query.
 ''';
   }